python_service/requirements.txt
# PyTorch CPU-only (lighter, no CUDA dependencies)
--index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=10.0.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0

